Extract fake events out.
